+++ packages/net-kit/lib/src/utility/file/file_reader.dart
@@ -1,0 +1,1 @@
+export 'file_reader_stub.dart' if (dart.library.io) 'file_reader_io.dart';